Move the kotlin code to the kotlin folder of the utils module #93
