+++ utils/src/main/kotlin/com/wansensoft/utils/Test.kt
@@ -10,7 +10,7 @@
  * OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
  * and limitations under the License.
  */
-package com.wansensoft.utils.enums;
+package com.wansensoft.utils;
 
 import lombok.Data
